rename `pluginservice.js` to `service.js`
