+++ index.js
@@ -2,7 +2,7 @@ import pluginConfig from './config';
 const {base, inherit} = g3wsdk.core.utils;
 const {Plugin} = g3wsdk.core.plugin;
 const {addI18nPlugin} = g3wsdk.core.i18n;
-const Service = require('./pluginservice');
+const Service = require('./service');
 
 const _Plugin = function() {
   base(this);
